requirements.txt
Flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
flask-limiter==3.5.0
Werkzeug==3.0.1 